requirements.txt
anyio==3.6.2
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
bcrypt==4.0.1
cffi==1.15.1
click==8.1.3
cryptography==39.0.0
decorator==5.1.1
executing==1.2.0
fastapi==0.74.1
h11==0.14.0
idna==3.4
ipdb==0.13.11
ipython==8.8.0
jedi==0.18.2
matplotlib-inline==0.1.6
parso==0.8.3
passlib==1.7.4
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.36
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==1.10.4
Pygments==2.14.0
PyJWT==2.6.0
python-multipart==0.0.5
six==1.16.0
sniffio==1.3.0
stack-data==0.6.2
starlette==0.17.1
tomli==2.0.1
traitlets==5.8.0
typing_extensions==4.4.0
uvicorn==0.17.5
wcwidth==0.2.5
